Implement new design for menu popup
We're using a custom PopupMenu implementation modified from the
appcompat library that supports a custom adapter. The appcompat
implementation is being used because it's internal implementation
classes are accessible, so we only need to copy the MenuPopupHelper
class in our codebase. We also have a basic level of custom dimentional
styles and attributes defined.