Fix issue with wrong LayoutParams in popup menu helper
ListView had a bug where it could not handle the base ViewGroup
LayoutParams in it's items without throwing a ClassCastException, that
was fixed in Lollipop. The generateLayoutParams() method definition in
ViewGroup that takes an existing base LayoutParams instance is
restricted to protected access level, while only the method that takes
an AttributeSet as parameter is publicly exposed. Therefore, in order to
generate the correct LayoutParams, we could either create an XML layout
consisting only of a bare View just to acquire the correct LayoutParams
in a generic manner, or hardcode it to use ListView.LayoutParams
explicitly. I opted to just replace the LayoutParams usage with setting
the minimum height instead, which works just as well, although it may
not be the semantically correct attribute to use.

A minor spelling error in the class header comment is also fixed.