Fix improper Phrase key formatting and usage in video deletion resource
Double curly braces were used for the key in the "other" plural string
resource, which is interpreted as an escape by Phrase and ignored.
Additionally, the video list module was using the wrong key for
formatting. Both of these things led to exceptions; in the recent videos
module there was a generic try catch block covering the whole cleanup,
and in the video list module there was no try catch block at all, which
caused some global exception handled to catch it. The try catch blocks
have been fixed to their proper scoped as well, to avoid a repeat of
this incidence.