Add IconImageView widget to android-iconify for RTL support in XML
RTL auto-mirroring support was added in IconDrawable in commit
3dcb06852a7d75b81643155cfc0695c5abc08c24. However, it is not supported
for TextView and it's subclasses, as they allow arbitrary text content.
Therefore, this commit adds an IconImageView widget with defined
attributes for icon and color. The icon attribute is of enum type, with
all the icons included, using the hash code of their names to map them
to the IconValue items.

Also added is a getIcon() method to the IconDrawable, in order to
facilitate the same method in IconImageView.