Fix null default value returned on database query exceptions
DbOperationBase catches any exceptions thrown during the query. If there
is a callback defined, then it's failure callback is triggered. However,
in the absence of a defined callback, the method synchronously returns
null, as the base method is not aware of the default value of the
subclass type. Catching the exception is thus rendered useless, as the
client code might as well catch the exception explicitly, as add null
checking on every database query. This commit adds an abstract method
for defining the default value, which will be returned instead of null
in the case of any exceptions.