Fix testing environment
RoboJuice integration in commit 5158557b9df55463301a58c3b186716b14cf330a
broke the testing environment, as it removed the mock Application and
Config implementations. It introduced a weak structure for setting up
the injections in tests, which is patched to work properly in this
commit. All the view related tests now extend HttpBaseTestCase so that
they can utilize the mock web server and have the other injections set
up as well.