Remove usage of newer API
Commit 465549909cabc3074c57007801b4f5d9ebbebb07 added validity checks
for the signup/login Activities using ISocial instances for
authentication. That check included checking for destruction using the
isDestroyed() method, which was introduced in API 17. This caused
devices with lower API levels to crash upon authentication.

This commit fixes this issue by removing the call to that method, and
using the lifecycle callbacks to set the Activity variable to null upon
destruction. This eliminates the need to use a weak reference as well,
so that has been removed as well. All the lifecycle callback methods are
now defined in the abstract ISocialImpl class, so that they don't need a
stub definition in all the implementations.