Fix IEdxEnvironment binding issue in tests caused by mocking
Commit 5158557b9df55463301a58c3b186716b14cf330a integrated RoboJuice,
but the mocking required for tests were implemented in an incomplete
manner. Among the API test bindings it defined was to bind an empty mock
of IEdxEnvironment to the interface. The bindings for the API test
suites were fixed in commit c60f898b8ef14ea745abc0d130faaf4691e1599d.
However, it also changed the structure to have all the view related
tests inherit from HttpBaseTestCase. This causes issues with their use
of the stub IEdxEnvironment binding. This commit removes this mock
binding and brings back the original bindings for IEdxEnvironment.