Remove redundant check for main thread on network operations
Commit 9b02037dddda805f71a794b6ce90d732d1416543 added a check for the
current thread in HttpRequestDelegate in order to verify that the main
thread was not being used for networking operations. A runtime exception
was thrown if that was the case.

However, this check is redundant since this is already handled by the
StrictMode thread policy, with an exception thrown by default when
targeting Honeycomb or higher. Moreover, since Robolectric currently
redirects all message loops to the main thread, this causes Robolectric
tests to fail.

This commit removes this check, and removes the code duplication in the
Robolectrie tests that were designed to avoid encountering this issue.