Add login in Robolectric UI testing setup
All of the UI that we test assumes that the user is logged in.
Therefore, this commit is designed to ensure that the login process is
called before all the UI tests. The login() method has been ported from
ApiTests to HttpBaseTestCase as a utility method, which is overriden
and annotated as a test case in ApiTests, and as a base dependency in
the UI test classes. There is a new base class called UiTest which
defines the dependency, that will be extended by all the UI test suites.