Fix Robolectrie tests started activities not cleared
The utility method for asserting the next started Activity in
BaseFragmentActivityTest was using peekNextStartedActivity(), which
caused the Intent chain to not be cleared, thus potentially causing
issues for other linked tests that expect a clear state. This commit
fixes this issue by changing the method to getNextStartedActivity(),
which has the side-effect of also clearing the records of that Activity
from the Intent log chain.