Add tests for navigating CourseUnitActivity
This commit adds tests for navigating from CourseDashboardActivity to
CourseOutlineActivity, from base CourseOutlineActivity to section
CourseOutlineActivity, and from CourseOutlineActivity to
CourseUnitNavigationActivity.

A shadow has been created for SafeAsyncTask in conformance with
the AsyncTask shadow to divert it's main executer to use Robolectric's
"background" thread scheduler, which is actually scheduled on the same
thread as the main events. To prevent potential deadlocking while
waiting for onPreCall() while the scheduler is paused,
AndroidCallableWrapper has been shadowed as well to call onPreCall()
directly from the "background" thread. This makes it easy to test the
navigation from CourseUnitNavigationActivity, which uses a subclass of
RoboAsyncTask (itself a subclass of SafeAsyncTask), to query the course
components data before starting the CourseOutlineActivity instance.

Since Robolectric doesn't actually start activities, it is only possible
to test forward navigation, and only up to one level. This bars us from
testing the back stack restructuring the we perform upon returning from
CourseUnitNavigationActivity after having switched to a unit from a
different section.