Fix type issue in CourseOutlineActivityTest
CourseOutlineActivity expect the course component ID delivered as a
String in it's Intent extras and saved state Bundle. However the test
suite was attempting to get it as a Serializable, even though it injects
it as String itself. This caused it to return null, which didn't cause
any issue in the test as it requested the data as Serializable from both
the original Bundle holding the expected values, and the Fragment Bundle
holding the values to be verified. Thus they were both null, and
therefore equal.