Assert that performClick() returns true in UI tests
I ignoring the result of performClick() in the UI tests on the
assumption that it was implemented to always return true. However, it's
documentation specifies that it returns true only if it has a click
listener assigned to it. This commit changes the tests using
performClick() to always assert that it returns true.