Fix wrong popup menu item view type count
Commit 67f17085fa68cee0791bcfb0fb7b7b4f136f10ed implemented a custom
implementation of PopupMenu supporting a custom adapter. Since there was
no simple way to set up padding in ListPopupMenu, the padding was
instead defined inside the adapter. The left/right padding was set on
each item in addition to any existing padding, and the top/bottom
padding was implemented as separate items consisting of empty Views.
However, this became problematic for accessibility, as the popup menu
reported two extra items than what were shown to the user.

This commit is designed to fix the accessibility issue by appending the
top/bottom padding on the background drawable instead, by using
LayerDrawable to add another layer to the drawable containing a
transparent ShapeDrawable with the desired padding applied, or setting
the ShapeDrawable directly as the background if there was no background
defined. This method was previously thought to be unoptimal, as it
diversifies the padding implementations and confuses the background
identity and detail, but now it remains as the only way to implement the
padding with proper accessibility.

The padding cannot be applied on the ListView directly, as ListPopupMenu
only initialized the ListView when it's about to display it, at which
time it also calculates the height from a combination of the height of
it's items and the top/bottom padding of it's background. Subsequent
changes to it's padding does not result in the height of the popup menu
being recalculated.

Note that the getItemViewType() method in the adapter was previously
incorrectly implemented to return 2, regardless of the fact that
(including the padding items) there were three types of items. This
didn't cause any issues in practice because our existing menus were not
large enough to ever require recycling the views. However, now after
removing the padding type, the returned value is now correct, and thus
this method does not need to be modified along with the other methods.