Handle case where local session cookie becomes invalid
Even if we have a cookie in the cookie store, it may not be valid. This
change ensures that we always have a valid cookie, by just fetching one
right before we load (this is a pretty unfortunate way to solve this
problem, since it involes making more requests than we would like). We
should probably not bother doing this if we *just* fetched a cookie.
Filed ?filter=1440://openedx.atlassian.net/browse/MA-1153 to follow up.

Also as that ticket says, in the long term we should adopt the new API
that lets you listen for actual HTTP responses.

JIRA: https://openedx.atlassian.net/browse/MA-1006