Fix reference to renamed method in merge
Commit d2bbfdaf3d370af52b3cc4573c17a7d224247a23 replaced the
"responsiveUI" block field with "multiDevice" and renamed the
isResponsiveIU() method to isMultiDevice() in CourseComponent. This
method was also called from the navigation testing that was added to
CourseUnitNavigationActivityTest in commit
028d99cea9ec57bb4fb5ab397f6b0b50d29c0dd7. However, during the merge,
the method call from the test was not changed. This commit fixes this
issue by changing the method name in the test.