Fix potential null reference access
Commit 9ad9df6cec4dc77a275ce50511c7e3d1fbdcb1d3 implemented
synchronization of the course unit navigation state to the outline back
stack upon returning to it. However the logic it used assumed that the
unit block path size would not be smaller than the size of the current
outline block path. Therefore it did not allow the possibility of null
when getting the block corresponding to the last block in the outline
path from the leaf. This commit fixes this issue by reverting the
equals() operation, and thus allowing the possibility of null as a
result of that check.