Fix popup animation tests
Commit d3d1909c81457fb7650fa9e0df4b00fc6bbd8ce0 changed the message bar
animation popup structure to use the animation end callback to hide the
View instead of posting a delayed Handler callback. Since Robolectric
has a custom handler for animations that processes them immediately,
this caused the check for the interim visibility of the View to fail, as
the state could not be accessed from the test.

This commit fixes the issue by pausing the foreground Scheduler so that
the interim state can be fixed. To accomplish this, the assert method
takes a Runnable as a parameter that triggers the animation, and takes
care of pausing and resetting the Scheduler at the appropriate times for
the test to work properly.

Also, that commit moved the animation definitions to XML, so the detail
verification (type, duration, fill-after) has been removed from the
test.