Replace mock config with real one in test
Commit 486bf4e7916bf2a2d091ced16111f80aee298401 used Mockito to mock a
Config instance used to test whether the social feature was enabled or
not. However, since Mockito does not perform Roboguice injections, this
was causing errors. This commit replaces the mocked Config instance with
the real Config (with a few fields mocked) that is properly initialized.
This was set up in BaseTestCase in commit
c60f898b8ef14ea745abc0d130faaf4691e1599d.