Fix options menu test for CourseBaseActivity
The test for verifying the state of the options menu in
CourseBaseActivity incorrectly simulated a click on the "view on web"
button out of it's initialized scope. Therefore it failed in screens
where the options menu was disabled, such as CourseDashboardActivity.
This commit fixes this issue by moving the click simulation in the
proper scope.