Fix custom content view usage in popup animation tests
The message bar animation popup tests replace the content view with a
custom view that is used in the test, as there is no fixed layout that
can be relied on. However, this causes issues with the RoboGuice
injections that were added in commit
5158557b9df55463301a58c3b186716b14cf330a, which throw exceptions when
not finding the expected views. This commit fixes the issue by only
adding the custom view to the content without replacing the existing
views. Also, the simulated Activity lifecycle has been advanced to use
the full set up in the tests, for better simulation of the actual
scenarios.