Fix fullscreen test on BaseFragmentActivity
Since the integration of RoboGuice in commit
5158557b9df55463301a58c3b186716b14cf330a, the full screen landscape
theme was being set only after the content (and decor) view had been set
up, and thus was never actually set up at all. As the theme change seems
to have been introduced for the benefit of CourseUnitNavigationActivity,
which was the only subclass to support landscape orientation, and it
handled hiding the action bar correctly by itself, this only had the
effect of causing the test cases to fail. This commit fixes that by
removing the needless (and potentially harmful) theme change altogether,
along with the test for it.