Fix course activity parameters in tests
Commit 8ddc43a5772698aa8149354689c286f1d77a7193 changed the parameters
provided to CourseUnitNavigationActivity to send the actual unit ID
instead of the ID of the base CourseComponent in the parent
CourseOutlineActivity, and rely upon it instead of sending the unit
separately. However, the tests for this navigation were not updated
correctly to reflect this. This commit fixes this issue in the tests.