Fix download event test for CourseOutlineActivity
CourseVideoListActivity overrides the method that is called upon
receiving the download event in CourseBaseActivity. However, the test
suite for CourseOutlineActivity (which is the only class the extends
CourseVideoListActivity) did not have the base test for the download
event overridden, thus causing it to fail. This commit fixes the issue
by overriding the test with assertions for the correct semantics.