Ignore abstract base UI test suite
The UiTest class was introduced in commit
67a80bc8026b51a81d54c6dbb6207399df4c8eee as an abstract super class for
all the UI test suites for the purpose of defining shared dependencies
for UI tests. However, since it inherits from HttpBaseTestCase and does
not explicitly defines itself to be ignored in the tests, the test
runner was reporting an error on it upon encountering no test methods in
it. This commit explicitly marks this class to be ignored in the tests
to fix this error.