Fix database sharing in download progress bar test
The download progress bar test for BaseFragmentActivity used database
entry manipulation to test the different states. The reference to the
database was retrieved through the DatabaseFactory, which created a new
instance of IDatabase for each caller. However, this was problematic for
Robolectrie tests, as it provides a random storage location that could
change unpredictably during the test. This commit fixes the issue by
injecting it via RoboGuice as a singleton, which provides the same
reference that is used in BaseFragmentActivity, thus ensuring that it
will only attempt to create it once.