Delegate async loading to appropriate Activities
When a user opened a module from CourseDashboardFragment, it used to
make an asynchronous query to load the data for the module before
starting it, blocking the user on the dashboard with a loading spinner.
This commit changes this behaviour by having the modules query their
data themselves.

The discussion module had a query for the course discussion info before
being started from the dashboard, but the data was not used at all (or
even passed to) the module, so the query has been removed.

The implementation of CourseBaseActivity and it's subclasses has been
changed to have the data loaded asynchronously. The null checks on the
course data have also been removed on the assumption that the course
module could never be started without valid course data.

As a result of this, CourseDashboardActivity and
CourseDetailInfoActivity have been changed to extend
BaseFragmentActivity instead of CourseBaseActivity. The only logic that
they took from CourseBaseActivity was the initialization and retention
of the course data, which has now been imported locally in these
classes.

The tests have been updated to reflect the inheritance change in
CourseDashboardActivity. CourseBaseActivity and CourseOutlineActivity
have also been updated to run both with and without the provided course
ID via parameterized testing, pending support for parameterized testing
in RobolectricGradleTestRunner.