Fix cookie refresh system for WebView
Since WebViewClient did not provide an error callback with the HTTP
error code prior to Marshmallow, commit
3ac13a7b92afe6ee468cc07b1efb7bf3f41ac295 changed the reauthentication
system to always reauthenticate and get a fresh cookie for each session
of CourseUnitWebviewFragment. This commit modifies it to assume that the
cookie will remain valid for at least an hour, and refresh based on the
freshness instead of on each session without a freshness duration. Also,
support has been added to use the error callback in Marshmallow to deal
with it based on the server response.