Fix RoboGuice injection issues
RoboGuice was integrated in commit
5158557b9df55463301a58c3b186716b14cf330a. However, some injections were
not set up properly. The null pointer exceptions that were caused by
this were swallowed by the assorted generic exception catching blocks
that are scattered throughout the codebase. This commit attempts to fix
these issues:

1) TranscriptDownloader did not set up RoboGuice to initialize it's
injection of the ServiceManager instance. This has been fixed. Related
to this, TranscriptManager has also been converted into a RoboGuice
singleton, and injected appropriately.

2) BrowserUtil was instantiated manually instead of being injected, and
thus it's injection of the IEdxEnvironment instance was not being
initialized as well. It has now been converted into a static utility
method, with the IEdxEnvironment instance injected as a static field
from our RoboGuice module.