Fix popup menu anchoring content area
ListPopupWindow automatically sets up it's vertical offset to align the
content area to the anchor instead of the entire window body, unless
it's defined as a non-zero value in the style. This commit fixes that by
dynamically setting the offset to be the value defined in the style,
thus marking the zero value as deliberate instead of unset.