Fix popup menu not supporting checkable behaviour
The custom popup menu implemented in commit
67f17085fa68cee0791bcfb0fb7b7b4f136f10ed copied the appcompat library
implementation, but reverted to using the framework MenuInflater on the
assumption that it provided the same results. However, due to some
internal implementation details, the MenuInflater only sets the
checkable behaviour defined in menu groups if we're using it's own
specific implementation of MenuItem. Therefore, this commit reverts back
to using the SupportMenuInflater from the appcompat library, in order to
support checkable behaviour.