Fix sort/filter popup layout in discussion threads
Previously, the popups were created using PopupMenu, which has a fixed
layout, and the icon was not implemented. This commit changes the
implementation to a Spinner with a custom popup style applied that makes
it use our popup background. Since the rows are now created dynamically
by the adapter, it's now changed to consist of just a TextView, with the
icon drawable applied directly to it. The sort icon was previously
constructed as a combination of two IconView items with the
"long_arrow_up" and "long_arrow_down" icons; this result is now
constructed into one drawable by using LayerDrawable with the
appropriate insets.

Since the popup overlapping is only supported post-Lollipop, it is
being manually backported by defining a vertical offset. This can be
removed when upgrading the appcompat library to version 23 or above,
which backports the property.