Fix sort/filter popup icon in discussion threads
Commit e32034044569d46b8c28e5c4cb27d5937375cdf7 changed IconDrawable to
provide a fake ConstantState that was actually bound to the drawable
instance, in order to make it work with LayerDrawable (used for
combining the up and down arrow icons to create the sort icon for
discussion threads), which used it to create a new instance of it's
children upon mutation regardless of whether they supported it.

This commit works around another bug in LayerDrawable before Lollipop,
where it didn't initialize the bounds or levels of it's children upon
mutation or cloning. The workaround is to copy these properties as part
of the ConstantState cloning mechanism in IconDrawable.