Remove generic popup background
Commit e32034044569d46b8c28e5c4cb27d5937375cdf7 set our custom popup
background as the value for the generic popup background theme
attribute. However, this is used in quite a few places by the framework
where this does not fit at all, such as the drag handle on editable text
selections. This commit fixes that by removing the background from the
generic theme attribute settings.