Hide menu on courseware until data load
Until the course data has been loaded, the filter menu is not useful
and the website link is not available, so there is no point in
displaying the menu. This also removes the need for a null check on the
course data in the "View on Web" menu.

The filtering menu initialization logic has been changed to use the
color selector resource as the icon color, to take advantage of the
support for color selectors in IconDrawable that was introduced in
commit f5a0609baf19f46141ff676a37fb6f44cb5ff602. Also, the filter menu
icon is now directly toggled upon selection, instead of causing the
whole menu to be rebuilt.

There are also small formatting and structural improvements in the menu
creation logic, and the per-second onTick() callback handling also been
optimized and refined. The tests have also been updated to reflect the
implementation changes.