Implement presenter for submenus in popup menu
The custom courseware popup menu implementation flattens top-level
submenus, but only added a presenter for the base builder, which caused
submenu customizations to not trigger a refresh. Since there is not
callback on submenu additions, we ensure the presenter for submenus in
the adapter callbacks, which are called right before display.