Restructure offline menu item handling
Previously, the offline menu item display handling was left to each
subclass implementation of BaseFragmentActivity. This commit moves this
repetitive logic inside the base class. It also fixes an issue with the
state not being updated on Activity resumption, by having
NetworkConnectivityReceiver post sticky events. Also, the unsafe global
static flag for indicating offline state is now removed, and references
are replaced with queries to the NetworkUtil#isConnected() utility
method. The test suites are also updated to reflect this.