Fix crash when inflating popup menu on API < 17
Granular padding support was introduced for the custom popup menu
implementation in commit 7c0da0014c4d2064717bdbc2c32bc44b464d0862.
However, it assumed that the "android:paddingStart" and
"android:paddingEnd" attributes that were introduced in Jelly Bean MR 1
would return 0 when queried for in older API levels, and arranged the
logic to fallback to left and right padding in that case. However, what
happens is that an exception is thrown when attempting to get a
dimension resource from these attributes. Also the fallback was
implemented individually for both start and end padding.

This commit fixes this issue by checking for the current SDK version,
and only querying the start and end padding attributed if supported.
Furthermore the fallback will now only occur if both start and end
padding attributes are undefined.