Fix courseware mode switcher menu icon not changing
Commit 2b2ab5055b3cc8f14d94b6e5a6f65d91fa76cb93 streamlined the
implementation of the courseware mode switcher menu, by attempting to
change the menu icon directly instead of invalidating the whole options
menu. However the implementation had an issue where the menu item
variable was overloaded by the popup menu item variable of the same name
in the parameter of the click listener anonymous inner class. This issue
is now fixed in this commit.