Display android-iconify icons in XML layout editor
Commit 68673e417a5e7a8fc66eca926ad8ee1d82090d9c updated the
android-iconify library that was included in the project, and exported
it as an external library project. However, since android-iconify v2
used pluggable modules for icon fonts, this caused the IconImageView and
IconTextView definitions in XML layouts to not be renderable by the
graphical editor, which only loads the necessary classes without
initializing the application itself. This commit fixes that by creating
IconImageViewXml and IconTextViewXml for XML definitions, which extend
the library widgets and load the Font Awesome module from their static
initializers. The EIconTextView widget has been changed to extend
IconTextViewXml. XML definitions should use these widgets to enable
graphical display, though the actual code should only reference the base
class types.

Furthermore, since the graphical editor seems to only load assets from
the immediate project, the Font Awesome TTF file has been also included
in the assets of the VideoLocker module. Since this will be the primary
asset that will be used in the compiled APK, it would always need to be
kept in sync with the version used in the module upon updating. The old
version of the font asset that was missed in the upgrade is now deleted
as well.