Document local Font Awesome asset
The Font Awesome TTF file was included in the local assets of the
VideoLocker module in commit 1f35f810e456bc1ca04a6f95d80df4dd317aea00,
in order to enable the graphical layout editor to render it. This commit
documents this in the dependency definition for the Font Awesome module
for android-iconify in build.gradle, along with the need to have it
kept in sync with module updates.