Fix popup positioning in discussion threads
Commit e32034044569d46b8c28e5c4cb27d5937375cdf7 implemented anchor
overlapping manually for the sort and filter popups, using vertical
offsets. This was meant to be reverted to use the appcompat library's
generic implementation upon upgrade to v23. However, this was not done
when it was upgraded in commit d1e86fb59dc4417c19edcf538d46fe56887a9fc7,
and since the anchor overlapping is enabled by default in the appcompat
library, this caused the negative offset to be applied twice pre-
Lollipop, resulting in the popup being anchored above the spinner view.
This commit fixes the issue.