Fix wrong inheritance syntax in popup menu style
The popup menu and list popup window styles were incorrectly using the
"android:parent" attribute for defining their inheritance, instead of
the correct "parent" tag without namespace, causing the base appcompat
attributes to not be applied. This issue is fixed by this commit. Also,
the unnecessary "@style/" prefix in the parents of these styles have
been removed for simplification.