Fix list item selector hardcoded white by default
Commit ea09d1884e30c8ca51a908de164c6c8e9372fcab standardized the list
view item selector across the whole app. However, the existing selector
unnecessarily defined a default state of white, which caused spinner
dropdowns with gray backgrounds to have items displayed with white
background upon being unpressed (if the finger is dragged away without
dismissing the dropdown). This commit fixes the issue by changing the
default state of the selector to transparent.