Remove incorrect annotation from android-iconify
The @CheckResult annotation was incorrectly applied on the setVisible()
method of the IconDrawable class in the android-iconify library, which
doesn't in fact require the caller to check the result for correct
implementation. This was causing the IconImageView and IconImageButton
implementations to be reported as an error. This commit fixes this issue
by removing the annotation.