Fix loading indicator layout in reset password dialog
The "Please Wait" text was not being vertically centered with the
loading ProgressBar. This commit fixes the layout to have vertically as
well as horizontally centered gravity. It also removes the redundunt
logic in the dialog to have the ProgressBar explicitly managed by the
task, since the whole layout needs to be managed. It wasn't doing
anything anyway, since the ProgressBar was retrieved with the wrong ID.
The ID has now been removed from the widget, as it should not be
explicitly managed.