Use Font Awesome spinner for indeterminate progress bar
IconProgressBar has been defined as a subclass of ProgressBar which
supports custom attributes for defining an android-iconify icon as the
indeterminate drawable. The standard theme attributes for ProgressBar
styles have been set to point to the styles for IconProgressBar, since
that should be used for all the progress bar definitions in XML layouts.

There are also two special progress bar layout resources defined with
layout attributes for direct inclusion in layouts via the include tag.
The main one is "loading_indicator" which is for centered gravity, and
the other is "button_progress_indicator" which is placed at the end of
button layouts. Both of them have visibility set to "gone" by default,
but the latter is defined to be visible in the graphical preview. Both
of them define a generic ID, and all the code referencing them has been
changed to reference these IDs.