Add progress indicator in discussions add buttons
Indeterminate progress indictors have been added to the end of the
buttons for adding discussion posts, responses, and comments. Since the
text still needs to be centered without reference to the indicator, this
necceciates using a FrameLayout containing the text view and progress
indicator widget. The existing style for the button has been unbundled
into two styles, for the layout and text view respectively. The same
thing has also been done for the login and register buttons.

The fragments for adding posts, comments, and responses have also had
their task initialization logic modified to not post error messages to
the activity background, which would not be visible in any case.