Fix issue with Android Studio not setting up app gradle plugin
When you first setup an app build the gradle plugin wasn't loading in
Android Studio and you'd get an error about the module not being
available. We load the plugin dynamically (since we just fetch it from
github and there's no nice easy way to do that) and the code to do that
(in build.gradle) was never getting executed.

This isn't a problem with command line builds, but it looks like android
studio is building things in a different order or processing
"settings.gradle" directly or something, so the code to load the plugin
in "buildSrc/build.gradle" wasn't ever triggered. This moves that code
to "settings.gradle" which seems to fix the problem.