Lock discussions add buttons while in progress
While DiscussionAddPostFragment disabled it's add button after usage
during the process of adding, this functionality was not implemented
for DiscussionAddResponseFragment and DiscussionAddCommentFragment. This
is now fixed by this commit. Also, the commented out code for calling
the hideProgress() method in DiscussionAddPostFragment has been removed,
as this logic has been refactored to BaseSingleFragmentActivity.