Change discussion API method naming pattern
The existing naming pattern always used the positive verb (e.g. 'vote'),
even though it accepts a boolean parameter. These are now prefixed with
'set' to be more descriptive. This affects the method names in
DiscussionAPI and DiscussionService, and the various Task subclasses
that invoke them.

The tasks have also been cleaned up in the following ways:

 - Formatting.
 - Object variable access changed from public to private final.
 - Unnecessary Boolean class definition for parameter and object
   variable changed to primitive boolean.
 - Explicit (and lenient) null check changed to @NonNull annotation on
   parameters.
 - Exception throwing declaration removed from call() method, since all
   exceptions are being caught.