Fix gray patch at the bottom of course images
Course images were being transformed top be anchored to the top of the
container while preserving aspect ration, as of commit
56706bb450cb4c037e04b848c8be2bcbadd7dbcf. This was accomplished by
creating a new Bitmap of the target dimensions, and painting the
original Bitmap on it's Canvas. However, it looks like the new Bitmap
has a black color by default, and there doesn't seem to be any simple
way to have it changed to transparent. Therefore, the Canvas is now
painted with white background, which matches the background of all the
screens where it is being used.