Simplify and fix videos module structure
The video module was implemented on the basis that PlayerFragment needed
to be set up with one retained instance. Since nested Fragments can't
have their instances retained, there was some convoluted and intricate
logic across both PlayerActivity and MyRecentVideosFragment designed to
have the PlayerFragment that was used by the latter to be set up at the
top level, while retaining the same behaviour as a nested Fragment.

However, the initial assumption that the PlayerFragment instance needed
to be retained is incorrect or obsolete, since it's already implemented
as a nested non-retained child in CourseUnitVideoFragment, where these
kinds of workarounds can't work since it's used in a ViewPager.
Therefore this commit removes the PlayerFragment instance retention, and
refactors it as nested in MyRecentVideosFragment and controlled by it's
lifecycle, which enables the MyVideosTabActivity logic to be simplified
to setting up a simple FragmentTabHost implementation.

The FragmentTabHost manages the tab Fragments in both portrait and
landscape modes, but the TabWidget is hidden in landscape since it's
only meant for fullscreen video. MyVideosTabActivity has screen
orientation locked by default as before, and it's the responsibility of
the tab Fragment (MyRecentVideosFragment in this case) to unlock it as
needed, and ensure that it's relocked on detachment.

This restructuring also fixes an issue with the previous implementation,
where MyRecentVideosFragment didn't retain the state of the
PlayerFragment upon returning from landscape to portrait mode. This was
caused by the fact that MyRecentVideosFragment detached and reattached
the PlayerFragment upon resuming in order to refresh the FragmentManager
restoration, since it doesn't restore to it's layout initially, as that
is not initialized at this point. This all is caused by the fact that
the PlayerFragment is added to the Activity's layout in landscape mode,
and to MyRecentVideosFragment's layout in portrait (using the Activity's
FragmentManager).

The callbacks on the PlayerFragment are now set as soon as it's variable
is initialized in MyRecentVideosFragment (i.e. when it's added or
restored). The listeners for the next and previous buttons are set up
when the PlayerFragment is initially shown, or alongside initializing
the list adapter data set if it's being restored.