Retain video model in PlayerFragment
PlayerFragment was storing the video model in a local variable, without
retaining it in its saved state, which was causing inconsistent state.
This isn't an issue in the current implementations as the only instances
of PlayerFragment in which the video model is queried externally are
retained, but it will cause issues when the implementation is changed.
This commit fixes the issue.